Refactor useQuote hook; isolate mapped attributes from API in specific DTO; remove useEffect to fetch quote on load from hook (move to specific page that requires it)
